Replace ConditionallyRender component with client component
